refactor: move populateMeta to helpers
