+++ packages/nocodb/src/lib/meta/api/helpers/index.ts
@@ -1,0 +1,3 @@
+import { populateMeta } from "./populateMeta";
+
+export { populateMeta }
